+++ prettier.js
@@ -1,5 +1,5 @@
 /**
- * This config should be used on top of any of the other configs, to disable Prettier handle rules.
+ * This config should be used on top of the other configs to disable Prettier handle rules.
  */
 export default [
   {